part_2/requirements.txt
fastapi 
"uvicorn[standard]" 
llama-index-core
llama-index-llms-openai 
openai 
pydantic-settings
# going to improve with pip-compile later so we lock versions used. See part 3